+++ src/features/basic/index.ts
@@ -14,6 +14,7 @@ import './clickable-apex-logo';
 import './co-base-count';
 import './contd-condition-address-placeholder';
 import './contd-upward-search-results';
+import './cx-search-bar';
 import './cxpo-auto-price';
 import './correct-commands/correct-commands';
 import './custom-item-sorting/custom-item-sorting';